Add color emoji support
This isn't full support. Emoji characters are stored as bitmaps. There won't always be a bitmap that is the perfect size so the nearest bitmap is found and that is used but a better solution is to find the nearest bitmap larger than necessary and then scale it down to the appropriate size.