Draft C++ Unpacker
The C implementation of unpacker worked.

TL;DR, Writing good C++ code is more important to me than writing bad cross-language C code.

I tested every feature and the whole thing worked fine. But I didn't like the code. I don't like C. Two of C's weaknesses resource management and error handling. If you want to write safe C then you'll have to write unmaintainable. When malloc fails, it returns NULL. So I have to check the pointer and make sure it's not NULL. If the pointer is NULL, I have to free all the resources I have allocated in this scope and return something that tells the caller that this function failed.

So you end up with a lot of duplication. A function is a great place to put duplicate code but you can't wrap this in a function. The function will have to return an error code and then that error code will have to be checked. I tried writing a macro that checked the pointer, freed resources and returned NULL. I'm not proud of that solution. It was probably better to just leave the function as it was.

C++ solves both of these problems: destructors and exceptions. When something goes wrong, you throw an exception and the destructors are automatically called. So you only have to check for errors if they aren't fatal. By that I mean, check that a memory allocation failed if you can do without that memory and continue.

Exceptions can actually generate faster code. Exceptions only incur a performance penalty if an exception is thrown. When checking the return value of a function in C, you have to pay for the branch every time the code is executed regardless of whether there was an error.